Move all pFq tests to a separate test file
