+++ ramanujantools/cmf/known_cmfs/known_cmfs_test.py
@@ -2,12 +2,10 @@ from pytest import approx
 
 import itertools
 
-import sympy as sp
-from sympy.abc import x, y, z, n
+from sympy.abc import x, y, n
 
-from ramanujantools import Matrix, IntegerRelation
 from ramanujantools.pcf import PCF
-from ramanujantools.cmf import known_cmfs, CMF
+from ramanujantools.cmf import known_cmfs
 
 
 def test_cmf_e():
@@ -54,171 +52,6 @@ def test_cmf1():
         assert limit.as_float() == approx(-a + b / limit.mp.log(1 + b / a), 1e-4)
 
 
-def test_2F1_theta_derivative():
-    x0 = sp.Symbol("x0")
-    x1 = sp.Symbol("x1")
-    y0 = sp.Symbol("y0")
-    z = sp.Symbol("z")
-    expected = CMF(
-        {
-            x0: Matrix(
-                [
-                    [1, -x1 * z / (z - 1)],
-                    [1 / x0, 1 - (x0 * z + x1 * z - y0 + 1) / (x0 * (z - 1))],
-                ]
-            ),
-            x1: Matrix(
-                [
-                    [1, -x0 * z / (z - 1)],
-                    [1 / x1, 1 - (x0 * z + x1 * z - y0 + 1) / (x1 * (z - 1))],
-                ]
-            ),
-            y0: Matrix(
-                [
-                    [
-                        y0 * (-x0 - x1 + y0) / (x0 * x1 - x0 * y0 - x1 * y0 + y0**2),
-                        x0 * x1 * y0 / (x0 * x1 - x0 * y0 - x1 * y0 + y0**2),
-                    ],
-                    [
-                        y0 * (1 - z) / (z * (x0 * x1 - x0 * y0 - x1 * y0 + y0**2)),
-                        y0**2 * (z - 1) / (z * (x0 * x1 - x0 * y0 - x1 * y0 + y0**2)),
-                    ],
-                ]
-            ),
-        }
-    )
-    cmf = known_cmfs.pFq(2, 1)
-    cmf.assert_conserving()
-    assert cmf == expected
-
-
-def test_2F1_theta_derivative_negate_denominator():
-    x0 = sp.Symbol("x0")
-    x1 = sp.Symbol("x1")
-    y0 = sp.Symbol("y0")
-    z = sp.Symbol("z")
-    expected = CMF(
-        {
-            x0: Matrix(
-                [
-                    [1, -x1 * z / (z - 1)],
-                    [1 / x0, 1 - (x0 * z + x1 * z + y0 + 1) / (x0 * (z - 1))],
-                ]
-            ),
-            x1: Matrix(
-                [
-                    [1, -x0 * z / (z - 1)],
-                    [1 / x1, 1 - (x0 * z + x1 * z + y0 + 1) / (x1 * (z - 1))],
-                ]
-            ),
-            y0: Matrix(
-                [
-                    [1, x0 * x1 * z / ((y0 + 1) * (z - 1))],
-                    [
-                        -1 / (y0 + 1),
-                        1 + (x0 * z + x1 * z + y0 + 1) / ((y0 + 1) * (z - 1)),
-                    ],
-                ]
-            ),
-        }
-    )
-    cmf = known_cmfs.pFq(2, 1, negate_denominator_params=True)
-    cmf.assert_conserving()
-    assert cmf == expected
-
-
-def test_2F1_normal_derivative():
-    x0 = sp.Symbol("x0")
-    x1 = sp.Symbol("x1")
-    y0 = sp.Symbol("y0")
-    z = sp.Symbol("z")
-    expected = CMF(
-        {
-            x0: Matrix(
-                [
-                    [1, -x1 / (z - 1)],
-                    [z / x0, 1 + (-x0 * z - x1 * z + y0 - 1) / (x0 * (z - 1))],
-                ]
-            ),
-            x1: Matrix(
-                [
-                    [1, -x0 / (z - 1)],
-                    [z / x1, 1 + (-x0 * z - x1 * z + y0 - 1) / (x1 * (z - 1))],
-                ]
-            ),
-            y0: Matrix(
-                [
-                    [
-                        y0 * (-x0 - x1 + y0) / (x0 * x1 - x0 * y0 - x1 * y0 + y0**2),
-                        x0 * x1 * y0 / (z * (x0 * x1 - x0 * y0 - x1 * y0 + y0**2)),
-                    ],
-                    [
-                        y0 * (1 - z) / (x0 * x1 - x0 * y0 - x1 * y0 + y0**2),
-                        y0**2 * (z - 1) / (z * (x0 * x1 - x0 * y0 - x1 * y0 + y0**2)),
-                    ],
-                ]
-            ),
-        }
-    )
-    cmf = known_cmfs.pFq(2, 1, theta_derivative=False)
-    cmf.assert_conserving()
-    assert cmf == expected
-
-
-def test_2F1_normal_derivative_negate_denominator():
-    x0 = sp.Symbol("x0")
-    x1 = sp.Symbol("x1")
-    y0 = sp.Symbol("y0")
-    z = sp.Symbol("z")
-    expected = CMF(
-        {
-            x0: Matrix(
-                [
-                    [1, -x1 / (z - 1)],
-                    [z / x0, 1 + (-x0 * z - x1 * z - y0 - 1) / (x0 * (z - 1))],
-                ]
-            ),
-            x1: Matrix(
-                [
-                    [1, -x0 / (z - 1)],
-                    [z / x1, 1 + (-x0 * z - x1 * z - y0 - 1) / (x1 * (z - 1))],
-                ]
-            ),
-            y0: Matrix(
-                [
-                    [1, x0 * x1 / ((y0 + 1) * (z - 1))],
-                    [
-                        -z / (y0 + 1),
-                        1 - (-x0 * z - x1 * z - y0 - 1) / ((y0 + 1) * (z - 1)),
-                    ],
-                ]
-            ),
-        }
-    )
-    cmf = known_cmfs.pFq(2, 1, theta_derivative=False, negate_denominator_params=True)
-    cmf.assert_conserving()
-    assert cmf == expected
-
-
-def test_2F1_z_evaluation():
-    p = 2
-    q = 1
-    z_value = -7
-    assert known_cmfs.pFq(p, q, z_eval=z_value) == known_cmfs.pFq(p, q).subs(
-        {z: z_value}
-    )
-
-
-def test_gamma():
-    cmf = known_cmfs.pFq(2, 2, negate_denominator_params=True, z_eval=-1)
-    x0, x1 = sp.symbols("x:2")
-    y0, y1 = sp.symbols("y:2")
-    trajectory = {x0: 1, x1: 1, y0: 1, y1: 0}
-    start = {x0: 1, x1: 1, y0: 1, y1: 1}
-    limit = cmf.limit(trajectory, 100, start)
-    assert IntegerRelation([[1, 3, 0], [-3, -5, 0]]) == limit.identify(limit.mp.euler)
-
-
 def test_all_conserving():
     r"""
     Checks that all cmfs are indeed cmfs.
@@ -236,7 +69,6 @@ def test_all_conserving():
     known_cmfs.cmf3_3().assert_conserving()
     known_cmfs.hypergeometric_derived_2F1().assert_conserving()
     known_cmfs.hypergeometric_derived_3F2().assert_conserving()
-    known_cmfs.pFq(2, 2).assert_conserving()  # randomly choosing 2F2
 
 
 def test_back_conserving():
