+++ ramanujantools/cmf/known_cmfs/pfq.py
@@ -4,6 +4,8 @@ from sympy.abc import z
 from ramanujantools import Matrix
 from ramanujantools.cmf import CMF
 
+from functools import lru_cache
+
 
 theta = sp.symbols("theta")
 
@@ -40,6 +42,27 @@ class pFq(CMF):
                 Otherwise, will use theta derivatives.
             negate_denominator_params: if set to True, will inverse all y matrices.
                 Otherwise, will substitute y with -y.
+        """
+        matrices = pFq.construct_matrices(
+            p, q, z_eval, theta_derivative, negate_denominator_params
+        )
+        self.p = p
+        self.q = q
+        self.theta_derivative = theta_derivative
+        self.negate_denominator_params = negate_denominator_params
+        super().__init__(matrices=matrices, validate=False)
+
+    @lru_cache
+    @staticmethod
+    def construct_matrices(
+        p: int,
+        q: int,
+        z_eval: sp.Expr = z,
+        theta_derivative: bool = True,
+        negate_denominator_params: bool = False,
+    ):
+        r"""
+        Constructs the pFq CMF matrices.
         """
         x = sp.symbols(f"x:{p}")
         y = sp.symbols(f"y:{q}")
@@ -90,4 +113,20 @@ class pFq(CMF):
 
         matrices = {x[i]: Matrix(M / x[i] + sp.eye(equation_size)) for i in range(p)}
         matrices.update(y_matrices)
-        super().__init__(matrices=matrices, validate=False)
+        return matrices
+
+    @staticmethod
+    def predict_N(p, q, z):
+        N = max(p, q + 1)
+        if z == 1 and p == q + 1:
+            N -= 1
+        return N
+
+    @staticmethod
+    def state_vector(p, q, z=z):
+        a_symbols = sp.symbols(f"a:{p}")
+        b_symbols = sp.symobls(f"b:{p}")
+        values = [sp.hyper(a_symbols, b_symbols, z)]
+        for _ in range(1, pFq.predict_N(p, q, z)):
+            values.append(sp.Derivative(values[-1], z))
+        return Matrix(values)
